change type float to integer

+++ db/migrate/20160627025530_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration
   def change
     create_table :products do |t|
       t.string :name
-      t.float :current_bit
+      t.integer :current_bit
       t.datetime :time_left
       t.references :user, index: true, foreign_key: true
